requirements.txt
pip==25.1.1
setuptools>=70.0.0
wheel>=0.44.0

# Web framework - production ready
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==22.0.0
pydantic
pydantic-settings

# AI & Computer Vision - optimized versions
google-generativeai==0.3.2
pillow
opencv-python-headless==4.10.0.84
scikit-image==0.24.0
scipy==1.14.1
piexif==1.1.3
imagehash==4.3.1
pytesseract==0.3.13
exifread==3.0.0
numpy==1.26.4

# HTTP & Async
httpx==0.26.0
aiofiles==23.2.1

# Firebase & Storage
firebase-admin==6.4.0
cloudinary==1.38.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
